Notification system implemented

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -79,20 +79,50 @@ export async function POST(
       );
     }
 
-    // If like model exists, don't create it
-    await prisma.like.upsert({
-      where: {
-        userId_postId: {
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
+      },
+    });
+
+    if (!post) {
+      return Response.json(
+        { error: "Post not found" },
+        {
+          status: 404,
+        },
+      );
+    }
+
+    // In prisma transactions, we pass array of queries and if one fail, other also won't execute
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      update: {},
-    });
+        update: {},
+      }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
     return new Response();
   } catch (error) {
@@ -122,13 +152,40 @@ export async function DELETE(
       );
     }
 
-    // delete can throw error if follower doesnt exist which can happen due to race conditions but deleteMany wont throw error
-    await prisma.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId,
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post) {
+      return Response.json(
+        { error: "Post not found" },
+        {
+          status: 404,
+        },
+      );
+    }
+
+    // delete can throw error if follower doesnt exist which can happen due to race conditions but deleteMany wont throw error
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      // Delete notification when a like on post is removed
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
